only destroy session if it exists
Reviewed-on: http://gerrit.tine20.com/customers/9235
Tested-by: Jenkins CI (http://ci.tine20.com/) <tine20-jenkins@metaways.de>

+++ tine20/Tinebase/Session/Abstract.php
@@ -99,7 +99,9 @@ abstract class Tinebase_Session_Abstract extends Zend_Session_Namespace
      */
     public static function destroyAndRemoveCookie()
     {
-        Zend_Session::destroy(true, true);
+        if (self::sessionExists()) {
+            Zend_Session::destroy(true, true);
+        }
     }
     
     /**
